requirements.txt
Flask
Pillow
torch
torchvision
Werkzeug
gunicorn
psutil 